Fixed Timing issues with "ENG - Shiguang Dailiren - ONA OP - Dive Back In Time"
I was working on this for a while, but never merged it in since I encountered an encoding issue when I tried extracting the OP. It seems like there's some sort of weird time issue. The time offset for my personal copy was closer than the copy in the repo, so I decided to combine the two.